// src/api/generated/model/commentsControllerGetComments200Item.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { CommentsControllerGetComments200ItemUser } from "./commentsControllerGetComments200ItemUser";

export type CommentsControllerGetComments200Item = {
  id?: string;
  content?: string;
  createdAt?: string;
  user?: CommentsControllerGetComments200ItemUser;
};